+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 ### Added
 
 ### Changed
+- Use string paths instead of FileInfo
 
 ### Deprecated